+++ src/app/core/@ngrx/router/router.actions.ts
@@ -1,0 +1,9 @@
+import { createAction, props } from '@ngrx/store';
+import { NavigationExtras } from '@angular/router';
+
+export const forward = createAction('[Router] FORWARD');
+export const back = createAction('[Router] BACK');
+export const go = createAction(
+  '[Router] GO',
+  props<{ path: any[]; queryParams?: object; extras?: NavigationExtras }>()
+);