+++ src/app/feature/admin/components/manage-products/manage-products.component.ts
@@ -5,6 +5,12 @@ import { Observable } from 'rxjs';
 import { CartService } from 'src/app/feature/cart/services/cart.service';
 import { ProductModel } from 'src/app/feature/product/models';
 import { ProductService } from 'src/app/feature/product/service/products.service';
+import { AppState, selectProducts } from 'src/app/core/@ngrx';
+
+import { Store, select } from '@ngrx/store';
+import { ProductsState } from 'src/app/core/@ngrx/products/products.state';
+import * as ProductActions from 'src/app/core/@ngrx/products/products.actions'
+import * as RouterActions from 'src/app/core/@ngrx/router/router.actions';
 
 @Component({
   selector: 'app-manage-products',
@@ -14,21 +20,35 @@ import { ProductService } from 'src/app/feature/product/service/products.service
 export class ManageProductsComponent implements OnInit {
 
   products$: Observable<ProductModel[]>;
+  productsState$: Observable<ProductsState>;
+  
   constructor(public productsService: ProductService,
-              private router: Router,
-              private cartService: CartService) { }
+    private router: Router,
+    private cartService: CartService,
+    private store: Store<AppState>) { }
 
   ngOnInit() {
-    this.products$ = this.productsService.getProducts();
+
+    console.log('We have a store! ', this.store);
+
+    this.products$ = this.store.pipe(select(selectProducts));
+    this.store.dispatch(ProductActions.getProducts());
+    //this.products$ = this.productsService.getProducts();//this.store.pipe(select([products]));//
 
   }
 
   onEdit(id: number) {
     const link = ['/admin/managingproducts/edit', id];
-    this.router.navigate(link);
+  
+    this.store.dispatch(RouterActions.go({
+      path: link,
+      queryParams:[{"id":id}]
+    }));
   }
   onAdd(id: number) {
     const link = ['/admin/managingproducts/add'];
-    this.router.navigate(link);
+    this.store.dispatch(RouterActions.go({
+      path: link
+    }));
   }
 }