+++ src/app/feature/product/models/index.ts
@@ -1,3 +1,4 @@
 
 export * from './product-feedback.model';
 export * from './product.model';
+export * from './product.interface';