+++ src/app/feature/product/models/product.interface.ts
@@ -1,8 +1,10 @@
 import { Size } from './Size.enum';
 
 export interface IProduct {
+     id:number,
      name: string;
      description: string;
      price: number;
-     category?: Size;
+     category: Size;
+     count: number
 }